backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
boto3==1.33.6
python-dotenv==1.0.0
